flask/requirements.txt
Flask
flask-cors
pytesseract
Pillow
google-generativeai